fix incompatible arg types throughout the code, especially around monitoring wrapper
The type of args is ambiguous within the DFK - someties its a list,
sometimes its a tuple... or at least it was at one point.

using a sequence for this also reflects that the args should not be
modified by the monitor wrapper

A function can be invoked with *args being any sequence, by the
looks of things, even though *args on the callee side will always
be a tuple - conversion to a sequence will happen as part of the
call, which is why this works.

is *args specifically a SEQUENCE, or some other sequence-like
type? would be good to annotate this in docstring of taskrecord
as a more principled argument rather than "lol we just change the
types every now and then"